Exclude issues with invalid project keys from all validation.
This fixes a bug in the previous implementation, whereby a commit would
be treated as valid due to validation not being performed on issues with
an invalid project key, but the list of matched issues being non-empty.

Since the list was non-empty, the check for *any* matched issue
succeeded, and the array of errors was empty: the commit was considered
valid.