Include the oauth setup URL in the error message
When the JIRA issue validation fails because there is no oauth token,
give the user the URL to use to create the trust relationship.

This is a much better user experience than telling the user to go and
find a commit in the web UI with a JIRA issue and click on it.